Added basic pages to be filled in
