+++ budget-planner/src/app/enter-receipt/page.tsx
@@ -1,0 +1,9 @@
+'use client'
+import React  from "react"
+
+export default class EnterReceipt extends React.Component {
+
+    render() {
+        return <p>Enter Receipt</p>
+    }
+}
