+++ budget-planner/src/app/transactions/page.tsx
@@ -1,0 +1,9 @@
+'use client'
+import React  from "react"
+
+export default class EnterReceipt extends React.Component {
+
+    render() {
+        return <p>Transactions</p>
+    }
+}
